+++ programs/drift/src/state/events.rs
@@ -442,6 +442,7 @@ pub struct LiquidationRecord {
     pub liquidate_perp_pnl_for_deposit: LiquidatePerpPnlForDepositRecord,
     pub perp_bankruptcy: PerpBankruptcyRecord,
     pub spot_bankruptcy: SpotBankruptcyRecord,
+    pub bit_flags: u8,
 }
 
 #[derive(Clone, Copy, AnchorSerialize, AnchorDeserialize, PartialEq, Eq, Default)]